Tons of changes read the description
Added Stylesheets to alot of the UI elements in the Wallet
Fixed any build errors/warnings during compile. Commented out an unused
function and added brackets around a specific if/else case.

To update the toolbar and make it look nicer, I had to create a QAction
in bitcoingui.h this specifically was a small button with no text as a
fake spacer type class I made.

Background color updates

Changed some of the defined colors to other things, (Unfirmed and unsync
colors from grey and red to white and orange.

Added a logo to the general overview ui page.

Renamed some UI elements, primarily on the Mining page ui

Updated the About page to include the dv8coin developers and a link to
the official dv8coin website.

Changed Dv8C to dv8c