Round of updates to makefile to get the windows
Round of updates to get the windows headless built, seems to run but
crashed after running a few commands such as getinfo, it did connect to
my other machine on testnet. I want to try a build without the extra
winsock include in compat.h, winsock2 should be used over this but I
added this trying to fix a build error and it is not needed.